pyproject.toml
[tool.poetry]
name = "pubmed_fetcher"
version = "0.1.0"
description = "Fetch research papers from PubMed based on a query."
authors = ["ruchir sarve <ruchirsarve@gmail.com>"]
packages = [{ include = "paper_fetcher" }]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.79"
python-dotenv = "^0.21.0"

[tool.poetry.scripts]
get-papers-list = "cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
